chamapro/requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
Django
django-allauth
django-crispy-forms
dj-database-url
gunicorn
idna
pillow
psycopg2-binary
python-decouple
python-dotenv
requests
sqlparse
tzdata
urllib3
whitenoise
